Make sure loans already invested into are properly ignored
First off, we need to figure out which loans cause blockations in the wallet. Then we need to add loans which are already covered. And finally, we need to send it all to the investing algorithm, so that it knows which loans to avoid.

Also, fix the investing URL so that it matches the current Zonky API.