Initial implementation of Zonkoid support (closes #72)
This commit introduces a generic mechanism through which remote services could be asked to provide a final decision on whether or not an investment should be made. There is one implementation of such service, Zonkoid - the support in Zonkoid mobile app is a work in progress.

Furthermore, strategies can now decide whether or not their investments require confirmation using the above mechanism.

And last but not least, because this is a major change, this commit increases mutation coverage to 85 % across al modules.