Create a new RESTEasy client for every request
RESTEasy clients seem to hold on to all the entities created through them. In order for this to not become a leak for long-term runs such as in daemon mode, we now create a new client every time a request is made, and close it immediately after the request has been fulfilled.