Refresh OAuth token even while sleeping (fixes #104)
This required changing quite a bit of the authenticating logic. Sleep now calls out to the AuthenticationHandler, which in turn can decide to forgo any authenticating if it decides no actual work can be performed with the authenticated APIs.

Some logging logic was also moved and removed, to make sure that, while sleeping, RoboZonky logs as little as possible.