Add support for universal pagination into the API (#124)
Several Zonky APIs support pagination. Until now, we've only supported that for BlockedAmount - and even then, the paging information had to be supplied directly as a RESTEasy client proxy method's argument.

This pull request changes that, allowing (as part of a internal API in robozonky-common) to read a collection of entities (currently Loan, Investment and BlockedAmount) as a Java Stream. This will in return automatically query the Zonky API as required to read all pages of the response.

This PR breaks a lot of the public APIs, but that is expected in a new major release.