Sacrifice HTTP redirect in order to properly enable socket timeouts
By debugging the RESTEasy code, we find out that the only way to
actually get it to honor the socket timeout is to have it completely
own the HTTP client. This pull request removes our own HTTP client,
capable of HTTP redirection, in favor of finally having RESTEasy
properly time out on requests.

This should allow us to get rid of the sudden death later.