Protect against failed portfolio updates (#185)
When a portfolio update fails, perhaps due to network connection loss, the robot used to treat it as an empty portfolio and would therefore over-invest certain ratings. With this commit, when portfolio update fails, no investments will be performed, as the failure will cause the daemon to fail and report such failure to the user.

In order for this situation to not be very frequent, this commit makes several more changes. Most notably, it now executes every single Zonky operation in its own call, guaranteeing the token will never go stale in the middle of a request.