Attempt to fix authentication problems
On slower or overloaded machines, pausing token refresh for a
long-running operation will result in self-inflicted damage.
The token will expire and no refresh will ever happen, because the
operation is preventing it.

We fix this in several ways. First, we never pause token refresh -
having the operation fail due to stale token and be immediately retried
is better than disabling the entire daemon. Second, certain operations
are now run in de-prioritized threads, so that daemon itself gets more
CPU time to run.