Isolate the disaster that is Refreshable
The wide-spread use of Refreshable was causing problems throughout the
code, and especially around token refresh. Refreshable was designed to
work on background without any external control - using it for tasks
that overlap with regular robot operation was a mistake, should never
have happened in the first place.

This commit removes Refreshable from everywhere except for the key use
cases - refreshing remote resources, such as configuration files. It is
also removed from ListenerService, where it was not necessary and was
easily replaced by a Supplier.

This change is so important for stability of the daemon that it warrants
breaking public APIs.