Pause the robot for important events
This is hopefully the final solution to the problem of internal data
updates possibly breaking the robot. It represents a major refactor of
the robot with the intention of making the internal dynamics and
dependencies of various threads more obvious and easier to maintain.

With this commit, robot is paused when it starts to update portfolio. If
the operation fails, robot remains paused and tries again and again,
using exponential backoff. After 1 hour, it gives up and terminates
itself. During this one hour, all other operations of the robot are
disabled.

This commit also introduces Zonky liveness check. If the API is not
available, be it for network reaosns or for Zonky downtime, all the
operations of the robot will be disabled until such a time when Zonky
can be reached again.