Reduce potential for Zonky token timing out (#236)
For some paginated reports, such as investment retrieval, the OAuth token was previously only ever requested before the first page. If the retrieval of pages took a long time, the token would expire in the meantime and the operation would always fail.

Therefore we refactor the code so that there is the option to refresh the token before every single Zonky API call. Rate limiting will be no match for us! :-)

During the work on this, Travis CI builds started failing due to a change in Oracle's Java deliverables. Therefore we also include change to the CI setup, or else the commit would never pass CI.