Implement Zonky transaction logging (fixes #240)
This is a new approach to tracking what is happening on Zonky servers. It reads the daily server update at 3am and then keeps a log of transaction made by both the robot and remotely. That log is then kept until such a time when a new remote update can be acquired, likely on the next day.

This commit also removes a lot of code that is no longer necessary in this new world.