Integrate with Stonky (closes #274) (#294)
Introduces a new Maven module for integration with Stonky. This integration allows to clone the master Google spreadsheet, check for updates and upload the Zonky exports. Furthermore, support is added to the CLI and to the Installer for setting up the Google credentials necessary to make this feature work. Some parts of existing code are refactored for better testability, in places where it clashed with this work.