Improve internal data consistency (#356)
Balance, blocked amounts and amounts at risk are now all retrieved and cached at the same time. Furthermore, balance and amounts are now based on the same synthetic charges. As a result, the data is as mutually consistent as humanly possible.

This change also allowed us to remove all the code related to portfolio update.