Add support for the reservation system (#389)
There are two regimes. In the first regime, the app only confirms reservations that match the existing strategy. In the second regime, the app also registers for every available queue, overriding any settings that the user may have previously kept.

This is a backwards incompatible change, requiring addition of some information into the existing strategies.