Only sell investments gradually not to skew the portfolio information (#414)
The goal of this is to only sell investments in small amounts, so that the entire portfolio is not destabilized.

This is necessary, since Zonky only recalculates their portfolios every 2 hours and investments sold between the recalculations therefore cause PortfolioOverview to become out of sync with reality. Since the selling operation is fundamentally asynchronous (we offer to sell, then someone in the future may decide to buy), we can not "fix" portfolio on our side, as that would require us to constantly query for sold investments, making sure we have included all of them in PortfolioOverview.

This will make sure we only sell so much at once as not to affect the portfolio too much.
SellingJob will ensure that the operation happens often enough that even large selloffs can be accomplished within days at worst.