Don't go into a death spiral on HTTP errors (#451)
Due to the recent surge in HTTP 429s, we are forced to introduce a better response handling mechanism.

The robot's behavior so far was to retry the operation when an error was received. However, when that error was "Too Many Requests" or a JSON parse error (due to our fault), this would only make the situation worse.

With this new behavior, the robot will pause itself if it receives any HTTP 5xx, HTTP 4xx or JSON parsing error. It will then perform exponential backoff until the operation no longer fails. This also replaces the existing cumbersome liveness checks.