Significantly improve performance of the core algorithms (#566)
Strategy now considers each loan/investment/participation individually, as opposed to together with every other. This means a lot of the operations that were being performed, such as ordering of the marketplace, are no longer performed, leading to considerable performance improvements.

One drawback of this new approach is that we are no longer able to prioritize assets that are relatively more underinvested than others. This feature is becoming less and less important, as it gets harder and harder to invest into something at all.